Refactor IUnityContainerConsumer as IUnityContainerServiceProvider.
