--- Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/Containers/IUnityContainerConsumer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b73fc56c50e68914cb5524072f27128f
