+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/Services/IUnityContainerServiceProvider.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 56d6d6ba57d1fbe4eb0e689de4c54d7a
